[Chore #164252950] put tests in a single suite
